refactor(github): extract renderers into src/gh/formatters.ts
Refs #801
